src/utils.py
import os
import yaml
from pathlib import Path

from dotenv import find_dotenv, load_dotenv

load_dotenv(find_dotenv())

PROJECT_ROOT = Path(__file__).resolve().parent.parent
DEFAULT_CONFIG_PATH = PROJECT_ROOT / "config.yaml"

def load_config():
    config_path = Path(os.getenv("CONFIG_PATH", DEFAULT_CONFIG_PATH))    
    if not config_path.exists():
        raise FileNotFoundError(f"Config file not found at {config_path}")

    with open(config_path, 'r') as file:
        return yaml.safe_load(file)


